backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
spacy
nltk
firebase-admin
pydantic
joblib
timexy